bug调用方法错误。

+++ dubbo-common/src/main/java/com/alibaba/dubbo/common/serialize/support/protostuff/ProtoOutput.java
@@ -239,7 +239,7 @@ public class ProtoOutput implements ObjectOutput {
                 writeFloat(v);
             } else if (obj instanceof Short) {
                 short v = (short) obj;
-                writeFloat(v);
+                writeShort(v);
             } else {
                 throw new RuntimeException("不支持的数字类型:" + obj.getClass().getName());
             }
